chore: 节点执行详情页重构 --ignore # Reviewed, transaction id: 3390
